// apps/recruiter-dashboard/__generated__/OrganizationSwitcherListFragment.graphql.ts
export type OrganizationSwitcherListFragment$data = {
  readonly id: string;
  readonly organizations: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly logoUrl: string;
        readonly name: string;
        readonly slug: string;
      };
    }>;
    readonly pageInfo: {
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "OrganizationSwitcherListFragment";
};
export type OrganizationSwitcherListFragment$key = {
  readonly " $data"?: OrganizationSwitcherListFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"OrganizationSwitcherListFragment">;
};